+++ internal/wingd/server.go
@@ -76,12 +76,14 @@ type Daemon struct {
 	appliedCores         float64
 	appliedMem           uint64
 
-	reservedCores float64
-	externalCores float64
-	reservedMem   uint64
-	externalMem   uint64
+	reservedCores      float64
+	externalCores      float64
+	externalAttributed bool
+	reservedMem        uint64
+	externalMem        uint64
 
-	attribution externalAttribution
+	attribution  externalAttribution
+	measuredPIDs map[int]struct{}
 
 	cpuMeasured bool
 	memMeasured bool
@@ -146,6 +148,7 @@ func New(cfg Config) (*Daemon, error) {
 		quit:                make(chan struct{}),
 		conns:               map[*conn]struct{}{},
 		byRun:               map[string]*conn{},
+		measuredPIDs:        map[int]struct{}{},
 		leaseRun:            map[admission.LeaseID]string{},
 		leaseCharge:         map[admission.LeaseID]wingwire.HostResources{},
 		leaseMembers:        map[admission.LeaseID][]string{},